requirements.txt
flask==3.0.0
anthropic==0.18.0
pydantic==2.5.0
python-dotenv==1.0.0
pillow==10.2.0